requirements.txt
hydra-core
pandas
pyarrow
scikit-learn
boto3